+++ src/main/java/com/example/security/service/ImageService.java
@@ -1,0 +1,6 @@
+package com.example.security.service;
+
+public interface ImageService {
+
+    int addImage(int id);
+}
